LayoutTests/js/dom/console-non-string-values-expected.txt

CONSOLE MESSAGE: line 9: I am a string
CONSOLE MESSAGE: line 10: 42
Should see two console messages, the first a string, and the second a number.
PASS successfullyParsed is true

TEST COMPLETE
